+++ include/nfx/Silicon.h
@@ -33,6 +33,7 @@
 // Gate primitives
 #include "nfx/silicon/gate/And.h"
 #include "nfx/silicon/gate/Nand.h"
+#include "nfx/silicon/gate/Nor.h"
 #include "nfx/silicon/gate/Not.h"
 #include "nfx/silicon/gate/Or.h"